Moved reloadTitle to convert-title.js from main. #28
